requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==3.2.23
django-cors-headers==3.7.0
djangorestframework==3.12.4
idna==3.10
openrouteservice==2.3.3
psycopg==3.2.3
pytz==2024.2
rapidfuzz==0.14.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
